# slices: return early in Repeat if count is 0