This change matches the implementation of strings/bytes.Repeat and
increases performance:

% benchstat before.out after.out    
goos: darwin
goarch: arm64
pkg: slices
cpu: Apple M1
              │  before.out  │              after.out               │
              │    sec/op    │    sec/op     vs base                │
Repeat/5x0-8    4.1840n ± 1%   0.4691n ± 0%  -88.79% (p=0.000 n=10)
Repeat/5x1-8     10.63n ± 1%    10.97n ± 0%   +3.15% (p=0.000 n=10)
Repeat/5x2-8     14.62n ± 0%    14.86n ± 0%   +1.61% (p=0.000 n=10)
Repeat/5x6-8     24.72n ± 0%    25.32n ± 0%   +2.43% (p=0.000 n=10)
Repeat/10x0-8   4.1915n ± 0%   0.4689n ± 0%  -88.81% (p=0.000 n=10)
Repeat/10x1-8    12.76n ± 0%    12.67n ± 0%   -0.67% (p=0.000 n=10)
Repeat/10x2-8    18.51n ± 0%    19.08n ± 0%   +3.08% (p=0.000 n=10)
Repeat/10x6-8    24.46n ± 0%    25.02n ± 0%   +2.31% (p=0.000 n=10)
geomean          11.85n         6.958n       -41.29%